Test computational complexity
Some bad news: after testing hundreds of playthroughs, I've found
that the average number of moves each turn on a h3 board is ~200
and on a h4 board it's ~1500. This includes opening/endgame turns
where there aren't many moves, so in reality the number can be as
high as ~3500 possible moves midgame. This means that the approach
I was taking for making a decent AI is all but useless. For now I
think I'll scrap the AI and come back when I have some more tools
at my disposal